Fix Java 9 warnings
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@1802297 13f79535-47bb-0310-9956-ffa450edef68

+++ batik-codec/src/main/java/org/apache/batik/ext/awt/image/codec/imageio/PNGTranscoderImageIOWriteAdapter.java
@@ -51,7 +51,7 @@ public class PNGTranscoderImageIOWriteAdapter implements
 
         int n = -1;
         if (hints.containsKey(PNGTranscoder.KEY_INDEXED)) {
-            n=((Integer)hints.get(PNGTranscoder.KEY_INDEXED)).intValue();
+            n= (Integer) hints.get(PNGTranscoder.KEY_INDEXED);
             if (n==1||n==2||n==4||n==8) 
                 //PNGEncodeParam.Palette can handle these numbers only.
                 img = IndexImage.getIndexedImage(img, 1<<n);
